pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis"
version = "0.1.0"
description = "AI orchestration system with a Chief of Staff managing expert agents"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "chromadb>=0.5.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0",
    "mcp[cli]>=1.26,<2",
    "pyobjc-framework-EventKit>=10.0; sys_platform == 'darwin'",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
include = ["memory*", "agents*", "documents*", "chief*", "tools*", "utils*", "apple_calendar*", "apple_notifications*", "apple_reminders*", "apple_mail*", "apple_messages*", "connectors*", "okr*"]

[tool.setuptools.package-data]
"agent_configs" = ["*.yaml"]

[tool.setuptools]
py-modules = ["mcp_server", "config"]

[project.scripts]
jarvis-mcp = "mcp_server:main"
